pyproject.toml
[project]
name = "python-telegrambot-petlin"
version = "0.1.0"
description = ""
authors = [
    {name = "Dmitry Petlin",email = "DimonRonD@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg2 (>=2.9.10,<3.0.0)",
    "pg (>=0.1,<0.2)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-telegram-bot (>=22.3,<23.0)",
    "django (>=5.2.5,<6.0.0)",
    "django-extensions (>=4.1,<5.0)"
]

[tool.poetry]
packages = [{include = "python_telegrambot_petlin", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
